chore: move musl to shared
